refactor: LazyColumn의 item을 컴포넌트로 분리
